requirements.txt
Flask
flask-cors
openai
gTTS
gunicorn
langdetect
google-api-python-client
google-auth